% Generated by roxygen2 (4.0.2): do not edit by hand
\name{contourLines}
\alias{contourLines}
\title{Obtain the contour lines of the space time plot.}
\usage{
contourLines(x)
}
\arguments{
\item{x}{A \emph{spaceTimePlot} object.}
}
\value{
Returns a matrix representing the contour lines of the
space time plot.
}
\description{
Obtain the contour lines of the space time plot.
}
\seealso{
\code{\link{spaceTimePlot}}
}

