% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corrDim.R
\name{corrMatrix}
\alias{corrMatrix}
\title{Returns the correlation sums stored in the \emph{corrDim} object}
\usage{
corrMatrix(x)
}
\arguments{
\item{x}{A \emph{corrDim} object.}
}
\value{
The \emph{corrMatrix} function returns the correlations matrix
storing the correlation sums that have been computed for all the embedding
dimensions.
}
\description{
Returns the correlation sums stored in the \emph{corrDim} object
}
\seealso{
\code{\link{corrDim}}
}

