% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFA.R
\name{fluctuationFunction}
\alias{fluctuationFunction}
\title{Returns the fluctuation function obtained in a DFA and represented by a
\emph{dfa} object.}
\usage{
fluctuationFunction(x)
}
\arguments{
\item{x}{A \emph{dfa} object.}
}
\value{
The \emph{fluctuationFunction} function returns the fluctuation 
function used obtained in the DFA.
}
\description{
Returns the fluctuation function obtained in a DFA and represented by a
\emph{dfa} object.
}
\seealso{
\code{\link{dfa}}
}
