% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceTimePlot.R
\name{getContourLines}
\alias{getContourLines}
\title{Obtain the contour lines of the space time plot.}
\usage{
getContourLines(x)
}
\arguments{
\item{x}{A \emph{spaceTimePlot} object.}
}
\description{
Obtain the contour lines of the space time plot.
}
\seealso{
\code{\link{spaceTimePlot}}
}
