% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearTests.R
\name{keenanTest}
\alias{keenanTest}
\title{Keenan's test}
\usage{
keenanTest(time.series, ...)
}
\arguments{
\item{time.series}{The original time.series.}

\item{...}{Additional arguments for the \code{\link[stats]{ar}} function.}
}
\value{
A list containing the results of the test, including:
\itemize{
   \item test.stat: the F-squared test statistic
   \item df1 and df2: the degrees of freedom of the test statistic.
   \item p.value.
   \item order: order of the AR process used for testing.
}
}
\description{
Keenan's test: test for nonlinearity against the null hypothesis that the time
series follows some AR process.
}
\references{
Keenan, D. M. (1985), A Tukey nonadditivity-type test for 
time series Nonlinearity, Biometrika, 72, 39-44.
}
\seealso{
\code{\link{nonlinearityTest}}, \code{\link{tsayTest}}, 
\code{\link{mcleodLiTest}},\code{\link{thresholdTest}}
}
