% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/informationDimension.R
\name{logRadius}
\alias{logRadius}
\title{Obtain the the average log(radius) computed
on the information dimension algorithm.}
\usage{
logRadius(x)
}
\arguments{
\item{x}{A \emph{infDim} object.}
}
\value{
A numeric vector representing the average log(radius) computed
on the information dimension algorithm represented by the \emph{infDim} object.
}
\description{
Obtain the the average log(radius) computed
on the information dimension algorithm.
}
\seealso{
\code{\link{infDim}}
}

