% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearTests.R
\name{mcleodLiTest}
\alias{mcleodLiTest}
\title{McLeod-Li test}
\usage{
mcleodLiTest(time.series, lag.max)
}
\arguments{
\item{time.series}{The original time.series.}

\item{lag.max}{Maximum number of lags for which to test for 
conditional heteroscedascity.}
}
\value{
A list containing the \emph{p.values} for each of the 
Ljung-Box tests computed using lags ranging from 1 to \emph{lag.max}.
}
\description{
McLeod-Li test for conditional heteroscedascity (ARCH).
}
\references{
Tsay, Ruey S., and Rong Chen. Nonlinear time series analysis. 
Vol. 891. John Wiley & Sons, 2018.
}
\seealso{
\code{\link{nonlinearityTest}}, \code{\link{keenanTest}},
 \code{\link{tsayTest}}, \code{\link{thresholdTest}}
}
