% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nlOrder}
\alias{nlOrder}
\title{Get the order of the nonlinear chaotic invariant.}
\usage{
nlOrder(x)
}
\arguments{
\item{x}{An object containing all the information needed for the estimate of 
the chaotic invariant.}
}
\value{
A numeric vector with the radius of the neighborhoods used for the 
computations of a chaotic invariant.
}
\description{
Get the order of the nonlinear chaotic invariant.
}
\references{
H. Kantz  and T. Schreiber: Nonlinear Time series Analysis 
(Cambridge university press)
}
\seealso{
\code{\link{corrDim}}, \code{\link{sampleEntropy}}
}
\author{
Constantino A. Garcia
}
