% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plotLocalScalingExp}
\alias{plotLocalScalingExp}
\title{Plot local scaling exponents}
\usage{
plotLocalScalingExp(x, ...)
}
\arguments{
\item{x}{An object containing all the information needed for the estimate of
the chaotic invariant.}

\item{...}{Additional graphical parameters.}
}
\description{
Plots the local scaling exponents of the correlation sum or 
the average Shannon  information (when computing information dimension).
}
\references{
H. Kantz  and T. Schreiber: Nonlinear Time series Analysis 
(Cambridge university press)
}
\author{
Constantino A. Garcia
}
