% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonlinearStatistics.R
\name{timeAsymmetry}
\alias{timeAsymmetry}
\title{Time Reversibility statistic}
\usage{
timeAsymmetry(time.series)
}
\arguments{
\item{time.series}{The time series used to compute the statistic.}
}
\value{
The time simmetry statistic.
}
\description{
Time Reversibility statistic
}
\details{
The time simmetry statistic  measures the asymmetry of a time series
under time reversal by calculating:
\deqn{E[s_n\cdot s_{n+1}^2] - E[s_n^2\cdot s_{n+1}] }{E[s_n * s_{n+1}^2] - E[s_n^2 * s_{n+1}] }.
Since linear stochastic series are symmetric under time reversal, this statistic
may be used for testing the assertion that the data was generated
from a stationary linear stochastic process or not.
}
\author{
Constantino A. Garcia
}

