% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearTests.R
\name{tsayTest}
\alias{tsayTest}
\title{Tsay's test}
\usage{
tsayTest(time.series, order)
}
\arguments{
\item{time.series}{The original time.series.}

\item{order}{Order used for the AR model.}
}
\value{
A list containing the results of the test, including:
\itemize{
   \item test.stat: the F-squared test statistic
   \item p.value.
   \item order: order of the AR process used for testing.
}
}
\description{
Tsay's test: test for nonlinearity against the null hypothesis that the time
series follows some AR process. This is a generalization of Keenan's test.
}
\references{
Tsay, R. S. (1986), Nonlinearity test for time series, Biometrika,
 73, 461-466.
}
\seealso{
\code{\link{nonlinearityTest}}, \code{\link{keenanTest}},
\code{\link{mcleodLiTest}},\code{\link{thresholdTest}}
}
