\name{windowSizes}
\alias{windowSizes}
\title{Returns the window sizes used for DFA in a \emph{dfa} object.}
\usage{
  windowSizes(x)
}
\arguments{
  \item{x}{A \emph{dfa} object.}
}
\value{
  The \emph{windowSizes} function returns the windows sizes
  used to detrend the time series in the DFA.
}
\description{
  Returns the window sizes used for DFA in a \emph{dfa}
  object.
}
\seealso{
  \code{\link{dfa}}
}

