% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lyapunov.R
\name{divTime}
\alias{divTime}
\title{Returns the time in which the divergence of close trajectories was computed 
in order to estimate the maximum Lyapunov exponent.}
\usage{
divTime(x)
}
\arguments{
\item{x}{A \emph{maxLyapunov} object.}
}
\value{
A numeric vector representing the time in which the divergence of
close trajectories was computed.
}
\description{
Returns the time in which the divergence of close trajectories was computed 
in order to estimate the maximum Lyapunov exponent.
}
\seealso{
\code{\link{maxLyapunov}}
}
