% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{embeddingDims}
\alias{embeddingDims}
\title{Get the embedding dimensions used for compute a chaotic invariant.}
\usage{
embeddingDims(x)
}
\arguments{
\item{x}{An object containing all the information needed for the estimate.}
}
\value{
A numeric vector with the embedding dimensions used for compute a
chaotic invariant.
}
\description{
Get the embedding dimensions used for compute a chaotic invariant.
}
\references{
H. Kantz  and T. Schreiber: Nonlinear Time series Analysis 
(Cambridge university press)
}
\author{
Constantino A. Garcia
}
