\name{sampleEntropyFunction}
\alias{sampleEntropyFunction}
\title{Returns the sample entropy function \eqn{h_q(m,r)} used for the computations
of the sample entropy.}
\usage{
  sampleEntropyFunction(x)
}
\arguments{
  \item{x}{A \emph{sampleEntropy} object.}
}
\value{
  A numeric matrix representing the sample entropy function
  \eqn{h_q(m,r)} obtained in #' the sample entropy
  computations represented by the \emph{sampleEntropy}
  object.
}
\description{
  Returns the sample entropy function \eqn{h_q(m,r)} used
  for the computations of the sample entropy.
}
\seealso{
  \code{\link{sampleEntropy}}
}

